--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,5 @@
 .stack-work/
+dist-newstyle/
 .DS_Store
 *~
 client_session_key.aes