--- a/hie.yaml
+++ b/hie.yaml
@@ -78,6 +78,12 @@ cradle:
     - path: "propagators/datadog/src"
       component: "hs-opentelemetry-propagator-datadog:lib"
 
+    - path: "propagators/datadog/test/spec"
+      component: "hs-opentelemetry-propagator-datadog:test:spec"
+
+    - path: "propagators/datadog/benchmark/header-codec/main.hs"
+      component: "hs-opentelemetry-propagator-datadog:bench:header-codec"
+
     - path: "propagators/w3c/src"
       component: "hs-opentelemetry-propagator-w3c:lib"