--- a/nix/sources.json
+++ b/nix/sources.json
@@ -5,22 +5,22 @@
         "homepage": "https://github.com/nmattia/niv",
         "owner": "nmattia",
         "repo": "niv",
-        "rev": "5830a4dd348d77e39a0f3c4c762ff2663b602d4c",
-        "sha256": "1d3lsrqvci4qz2hwjrcnd8h5vfkg8aypq3sjd4g3izbc8frwz5sm",
+        "rev": "82e5cd1ad3c387863f0545d7591512e76ab0fc41",
+        "sha256": "090l219mzc0gi33i3psgph6s2pwsc8qy4lyrqjdj4qzkvmaj65a7",
         "type": "tarball",
-        "url": "https://github.com/nmattia/niv/archive/5830a4dd348d77e39a0f3c4c762ff2663b602d4c.tar.gz",
+        "url": "https://github.com/nmattia/niv/archive/82e5cd1ad3c387863f0545d7591512e76ab0fc41.tar.gz",
         "url_template": "https://github.com/<owner>/<repo>/archive/<rev>.tar.gz"
     },
     "nixpkgs": {
-        "branch": "release-21.11",
+        "branch": "release-22.05",
         "description": "Nix Packages collection",
         "homepage": "",
         "owner": "NixOS",
         "repo": "nixpkgs",
-        "rev": "fe6f909f7d63fc916490e4a487f5bc60f36dabaf",
-        "sha256": "06wjbv1y00hxl00dnqa2lavrbwbid7ms45xijnc79mqnawbw0gz0",
+        "rev": "c8e459ac45ecdfd0042ae79fb1a115811f50021e",
+        "sha256": "0nrvckpi4jj86mpwhwyinmcq0zri9xmrdrf1q3g7ky5c3zlxz0wh",
         "type": "tarball",
-        "url": "https://github.com/NixOS/nixpkgs/archive/fe6f909f7d63fc916490e4a487f5bc60f36dabaf.tar.gz",
+        "url": "https://github.com/NixOS/nixpkgs/archive/c8e459ac45ecdfd0042ae79fb1a115811f50021e.tar.gz",
         "url_template": "https://github.com/<owner>/<repo>/archive/<rev>.tar.gz"
     }
 }