--- a/propagators/datadog/hs-opentelemetry-propagator-datadog.cabal
+++ b/propagators/datadog/hs-opentelemetry-propagator-datadog.cabal
@@ -5,18 +5,25 @@ version: 0.0.0.0
 author: Kazuki Okamoto (岡本和樹)
 maintainer: kazuki.okamoto@herp.co.jp
 
+common common
+  build-depends: base >= 4 && < 5
+  ghc-options: -Wall
+  if impl(ghc >= 8.0)
+    ghc-options: -Wcompat
+  default-language: Haskell2010
+
 library
+  import: common
   hs-source-dirs: src
   exposed-modules: OpenTelemetry.Propagator.Datadog
-  build-depends: base == 4.*,
-                 bytestring,
+                   OpenTelemetry.Propagator.Datadog.Internal
+  build-depends: bytestring,
                  hs-opentelemetry-api,
                  hs-opentelemetry-sdk,
-                 http-client,
                  http-types,
+                 primitive,
                  text
-  ghc-options: -Wall
-               -Wcompat
+  ghc-options: -Wcompat
                -Wno-name-shadowing
   if impl(ghc >= 6.4)
     ghc-options: -Wincomplete-record-updates
@@ -45,4 +52,32 @@ library
     ghc-options: -Wmissing-kind-signatures
                  -Woperator-whitespace
                  -Wredundant-bang-patterns
-  default-language: Haskell2010
+
+test-suite spec
+  import: common
+  type: exitcode-stdio-1.0
+  main-is: Spec.hs
+  hs-source-dirs: test/spec
+  other-modules: OpenTelemetry.Propagator.Datadog.InternalSpec
+  build-depends: hs-opentelemetry-propagator-datadog,
+                 hs-opentelemetry-api,
+                 bytestring,
+                 hspec,
+                 pretty-hex,
+                 QuickCheck
+  ghc-options: -threaded
+               -rtsopts
+               -with-rtsopts=-N
+  build-tool-depends: hspec-discover:hspec-discover
+
+benchmark header-codec
+  import: common
+  type: exitcode-stdio-1.0
+  main-is: main.hs
+  other-modules:
+  hs-source-dirs: benchmark/header-codec
+  build-depends: hs-opentelemetry-propagator-datadog,
+                 hs-opentelemetry-api,
+                 bytestring,
+                 criterion,
+                 deepseq