--- a/propagators/datadog/test/spec/Spec.hs
+++ b/propagators/datadog/test/spec/Spec.hs
@@ -1,0 +1,1 @@
+{-# OPTIONS_GHC -F -pgmF hspec-discover #-}