```
benchmarking newTraceIdFromHeader/new
time                 12.44 ns   (12.03 ns .. 12.95 ns)
                     0.994 R²   (0.991 R² .. 1.000 R²)
mean                 12.19 ns   (12.06 ns .. 12.47 ns)
std dev              607.9 ps   (352.0 ps .. 939.6 ps)
variance introduced by outliers: 74% (severely inflated)

benchmarking newTraceIdFromHeader/old
time                 616.2 ns   (610.6 ns .. 624.3 ns)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 621.2 ns   (615.7 ns .. 628.9 ns)
std dev              21.93 ns   (16.08 ns .. 27.70 ns)
variance introduced by outliers: 51% (severely inflated)

benchmarking newHeaderFromTraceId/new
time                 100.4 ns   (99.75 ns .. 101.3 ns)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 100.4 ns   (99.97 ns .. 101.1 ns)
std dev              1.790 ns   (1.204 ns .. 2.637 ns)
variance introduced by outliers: 23% (moderately inflated)

benchmarking newHeaderFromTraceId/old
time                 175.7 ns   (174.9 ns .. 177.0 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 175.5 ns   (175.1 ns .. 176.2 ns)
std dev              1.902 ns   (1.455 ns .. 2.570 ns)
```