@ruicc 2022-08-22
quickcheckでテストが書かれてるのは良いですね